+++ objectbox/lib/src/native/box.dart
@@ -346,23 +346,21 @@ class Box<T> {
   }
 
   void _putToOneRelFields(T object, PutMode mode, Transaction tx) {
-    _entity.toOneRelations(object).forEach((ToOne rel) {
-      if (!rel.hasValue) return;
-      rel.attach(_store);
-      // put new objects
-      if (rel.targetId == 0) {
-        rel.targetId =
-            InternalToOneAccess.targetBox(rel)._put(rel.target, mode, tx);
-      }
-    });
+    for (var toOne in _entity.toOneRelations(object)) {
+      // To avoid all ToOnes obtaining a Store for each put,
+      // pass the store of this box.
+      toOne.applyToDb(_store, mode, tx);
+    }
   }
 
   void _putToManyRelFields(T object, PutMode mode, Transaction tx) {
     _entity.toManyRelations(object).forEach((RelInfo info, ToMany rel) {
       // Always set relation info so ToMany applyToDb can be used after initial put
-      InternalToManyAccess.setRelInfo(rel, _store, info, this);
+      InternalToManyAccess.setRelInfo<T>(rel, _store, info);
       if (InternalToManyAccess.hasPendingDbChanges(rel)) {
-        rel.applyToDb(mode: mode, tx: tx);
+        // To avoid all ToManys obtaining a Store for each put,
+        // pass the store of this box.
+        rel.applyToDb(existingStore: _store, mode: mode, tx: tx);
       }
     });
   }