+++ objectbox/lib/src/native/store.dart
@@ -22,8 +22,10 @@ import 'bindings/helpers.dart';
 import 'box.dart';
 import 'model.dart';
 import 'sync.dart';
+import 'weak_store.dart';
 
 part 'observable.dart';
+part 'store_config.dart';
 
 /// Represents an ObjectBox database and works together with [Box] to allow
 /// getting and putting.
@@ -35,6 +37,7 @@ class Store {
   /// This meant for tests only; do not enable for releases!
   static bool debugLogs = false;
 
+  /// Pointer to the C instance of this, access via [_ptr] with closed check.
   late Pointer<OBX_store> _cStore;
   late final Pointer<OBX_dart_finalizer> _cFinalizer;
   HashMap<int, Type>? _entityTypeById;
@@ -57,9 +60,13 @@ class Store {
   /// A list of observers of the Store.close() event.
   final _onClose = <dynamic, void Function()>{};
 
-  /// If weak and calling [close] does not try to close the native Store and
+  /// If true and calling [close] will also close the native Store and
   /// remove [_absoluteDirectoryPath] from [_openStoreDirectories].
-  final bool _weak;
+  final bool _closesNativeStore;
+
+  /// If true and calling [close] will also close a cached WeakStore, if one was
+  /// created for this store.
+  final bool _closesWeakStore;
 
   /// Default value for string query conditions [caseSensitive] argument.
   final bool _queriesCaseSensitiveDefault;
@@ -150,7 +157,8 @@ class Store {
       bool queriesCaseSensitiveDefault = true,
       String? macosApplicationGroup})
       : _defs = modelDefinition,
-        _weak = false,
+        _closesNativeStore = true,
+        _closesWeakStore = true,
         _queriesCaseSensitiveDefault = queriesCaseSensitiveDefault,
         directoryPath = _safeDirectoryPath(directory),
         _absoluteDirectoryPath =
@@ -266,8 +274,10 @@ class Store {
   /// ```
   Store.fromReference(this._defs, this._reference,
       {bool queriesCaseSensitiveDefault = true})
-      // must not close the same native store twice so [_weak]=true
-      : _weak = true,
+      :
+        // Must not close native store twice, only original store is allowed to.
+        _closesNativeStore = false,
+        _closesWeakStore = true,
         directoryPath = '',
         _absoluteDirectoryPath = '',
         _queriesCaseSensitiveDefault = queriesCaseSensitiveDefault {
@@ -294,7 +304,9 @@ class Store {
   /// See [_clone] for details.
   Store._minimal(int ptrAddress, {bool queriesCaseSensitiveDefault = true})
       : _defs = null,
-        _weak = false,
+        _closesNativeStore = true,
+        // Minimal store does not have a model, so could not create Weak Store.
+        _closesWeakStore = false,
         directoryPath = '',
         _absoluteDirectoryPath = '',
         _queriesCaseSensitiveDefault = queriesCaseSensitiveDefault {
@@ -319,8 +331,8 @@ class Store {
   /// is closed (e.g. when the app exits).
   Store.attach(this._defs, String? directoryPath,
       {bool queriesCaseSensitiveDefault = true})
-      // _weak = false so store can be closed.
-      : _weak = false,
+      : _closesNativeStore = true,
+        _closesWeakStore = true,
         _queriesCaseSensitiveDefault = queriesCaseSensitiveDefault,
         directoryPath = _safeDirectoryPath(directoryPath),
         _absoluteDirectoryPath =
@@ -355,6 +367,27 @@ class Store {
       _reader.clear();
       rethrow;
     }
+  }
+
+  /// Obtains a Store from a weak Store reference for short-time use.
+  ///
+  /// Will throw if the Store is already closed.
+  /// Make sure to [close] this when done using, which will not close the
+  /// underlying store.
+  Store._fromWeakStore(
+      StoreConfiguration configuration, Pointer<OBX_weak_store> weakStorePtr)
+      : _defs = configuration.modelDefinition,
+        _closesNativeStore = true,
+        // This Store was obtained from a WeakStore, do not close it.
+        _closesWeakStore = false,
+        directoryPath = configuration.directoryPath,
+        _absoluteDirectoryPath = '',
+        _queriesCaseSensitiveDefault =
+            configuration.queriesCaseSensitiveDefault {
+    Pointer<OBX_store>? storePtr = C.weak_store_lock(weakStorePtr);
+    _checkStorePointer(storePtr);
+    _cStore = storePtr;
+    _attachFinalizer();
   }
 
   void _checkStoreDirectoryNotOpen() {
@@ -477,7 +510,12 @@ class Store {
 
     _reader.clear();
 
-    if (!_weak) {
+    if (_closesWeakStore) {
+      // If there is a weak store reference for this, close it as well.
+      WeakStore.get(configuration())?.close();
+    }
+
+    if (_closesNativeStore) {
       _openStoreDirectories.remove(_absoluteDirectoryPath);
       final errors = List.filled(2, 0);
       if (_cFinalizer != nullptr) {
@@ -766,6 +804,24 @@ class Store {
     final model = _defs;
     if (model == null) throw StateError('Minimal store does not have a model');
     return model;
+  }
+}
+
+/// This hides away methods from the public API
+/// (this is not marked as show in objectbox.dart)
+/// while remaining accessible by other libraries in this package.
+extension StoreInternal on Store {
+  /// See [Store._fromWeakStore].
+  static Store fromWeakStore(StoreConfiguration configuration,
+          Pointer<OBX_weak_store> weakStorePtr) =>
+      Store._fromWeakStore(configuration, weakStorePtr);
+
+  /// Returns the configuration for this to be used with
+  /// [WeakStore], valid for the lifetime of the process.
+  StoreConfiguration configuration() {
+    int id = C.store_id(_ptr);
+    return StoreConfiguration(
+        _modelDefinition, directoryPath, _queriesCaseSensitiveDefault, id);
   }
 }